One pull request to (hopefully) clear the board on the problems Matt and his team found (#36)
Close #33 
Close #32 
Close #31 
Close #28 
Ready for Laravel 8 once dependencies I don't control are updated.

+++ src/ForgetSheet.php
@@ -3,13 +3,34 @@
 namespace Grosv\EloquentSheets;
 
 use Illuminate\Support\Facades\File;
+use Illuminate\Support\Str;
 
 class ForgetSheet
 {
     public function execute($id)
     {
-        File::delete(config('sushi.cache-path').'/'.$id.'.sqlite');
+        if (File::isFile(config('sushi.cache-path').'/'.$id.'.sqlite')) {
+            File::delete(config('sushi.cache-path').'/'.$id.'.sqlite');
+        }
+        if (File::isFile(storage_path('framework/cache/'.$id.'.sqlite'))) {
+            File::delete(storage_path('framework/cache/'.$id.'.sqlite'));
+        }
+
+        $this->triggerCallback($id);
 
         return response()->noContent();
+    }
+
+    protected function triggerCallback($id)
+    {
+        $class = 'App\\'.Str::studly(str_replace('sushi-', '', $id));
+        if (! class_exists($class)) {
+            return;
+        }
+        $model = new $class();
+
+        if (method_exists($model, 'onForget')) {
+            $model->onForget();
+        }
     }
 }
